+++ src/constants/ActionTypes.js
@@ -1,9 +1,3 @@
-export default {
-	// Counter
-	INCREMENT_COUNTER: 'INCREMENT_COUNTER',
-	DECREMENT_COUNTER: 'DECREMENT_COUNTER',
-};
-
 export const LOAD_MESSAGES_SUCCESS = 'LOAD_MESSAGES_SUCCESS';
 export const LOAD_MESSAGES_ERROR = 'LOAD_MESSAGES_ERROR';
 export const LOAD_CHATS_SUCCESS = 'LOAD_CHATS_SUCESS';
