+++ src/tests/HeaderChat.test.js
@@ -4,7 +4,7 @@ import { Provider } from 'react-redux';
 import { MemoryRouter } from 'react-router-dom';
 import renderer from 'react-test-renderer';
 import configureStore from 'redux-mock-store';
-import HeaderChat from '../components/chatform/ChatHeaderInfo';
+import HeaderChat from '../components/chatform/HeaderChat';
 
 const mockStore = configureStore();
