+++ src/tests/HeaderTop.test.js
@@ -1,6 +1,6 @@
 import React from 'react';  
 import renderer from 'react-test-renderer'; 
-import { HeaderTop } from '../components/chatslist/ListOfChatsHeader';  
+import { HeaderTop } from '../components/chatslist/HeaderTop';  
 
 
 it('HeaderTop renders correctly', () => {
