+++ src/tests/PersonalFields.test.js
@@ -4,7 +4,7 @@ import { Provider } from 'react-redux';
 import { MemoryRouter } from 'react-router-dom';
 import renderer from 'react-test-renderer';
 import configureStore from 'redux-mock-store';
-import PersonalFields from '../components/userpage/UserInfoFields';
+import PersonalFields from '../components/userpage/PersonalFields';
 
 const mockStore = configureStore();
